requirements.txt
fastapi==0.111.0
psycopg2==2.9.9
PyJWT==2.8.0
pandas==2.2.2
SQLAlchemy==2.0.31
pytest==8.1.1
httpx==0.27.0
uvicorn
python-dotenv
